Remove the user.list for LDAP users.
Seperate the userlist of LDAP users with password user list so that
we do not need to add username in user.txt file for LDAP users. This
will benefit from the pg-ldap-sync project which provide a way to
sync user role of LDAP automatically.
When all LDAP users use the same hba rule, no changes will be made
if we use pg-ldap-sync, user roles will be created or dropped
according to the LDAP server when customer uses pgbouncer through
LDAP authentication.

After this commits, there will be only two configs that should be
changed if users wants to login through LDAP authentication:
1. auth_hba_file
    Example:
    `cat hba.conf`
    `host all user1 0.0.0.0/0 ldap ldapserver=10.117.190.215 ldapbasedn="CN=Users,DC=greenplum,DC=org" ldapbinddn="CN=Administrator,CN=Users,DC=greenplum,DC=org" ldapbindpasswd="Welcome1!" ldapsearchattribute="sAMAccountname"`

2. auth_type
    `auth_type = hba`