Allow admin_user to login to pgbouncer database when auth_type=hba.
As for LDAP, the auth_type has to be set to 'hba'. There will be a
side effect that admin_users cannot login into admin database pgobuncer.
So this commit remove the limitation of 'auth_type = any' to allow
admin user login into pgbouncer database.

In this case, there should be one entry for user pgbouncer in hba.conf:
`host all pgbouncer 0.0.0.0/0 trust`