Keep md5 and ldap user in the same tree. (#23)
Some customers have the need to switch authentication method
by changing the auth_hba_file and reload without restarting pgbouncer.
However the ldap implementation adds new users in a different user tree.
So when the customer swtiches the authentication method, it tries
to search a created pool on the new user->pools list. As the new user
and the old user have the same username, but do not point to the same
structure, so it will try to add a new pool to pool_list which
will trigger a fatal error because a pool with the same username and
dbname has already been added in pool_list.

This commmit will remove the difference between md5 and ldap user so
that the old and new user could share the same structure. So there is
no need to add a new pool to pool_list when customer switch the
authentication method because the username and dbname have not changed.